Move impl FromStr for Netv6Addr into the netv6addr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
